init/u-boot: Found root=
FROM:

    [    1.072421]  __irq_svc from default_idle_call+0x2c/0x30
    [    1.072439]  default_idle_call from do_idle+0xa4/0x118
    [    1.072458]  do_idle from cpu_startup_entry+0x18/0x1c
    [    1.072473]  cpu_startup_entry from rest_init+0xac/0xb0
    [    1.072511]  rest_init from arch_post_acpi_subsys_init+0x0/0x8
    [    1.079565] ---[ end Kernel panic - not syncing: VFS: Unable to mount root fs on unknown-block(0,0) ]---

TO:

    [    1.566641] 1f80: 00000002 c0d04f14 c10b8840 c0d04f58 4000406a 410fc075 00000000 00000000
    [    1.566651] 1fa0: 00000000 f0861fc0 c087308c c087345c 60000013 ffffffff
    [    1.566659]  __irq_svc from default_idle_call+0x2c/0x30
    [    1.566681]  default_idle_call from do_idle+0xa4/0x118
    [    1.566702]  do_idle from cpu_startup_entry+0x18/0x1c
    [    1.566720]  cpu_startup_entry from secondary_start_kernel+0x118/0x120
    [    1.566737]  secondary_start_kernel from 0x40101460
    [    1.574161] ---[ end Kernel panic - not syncing: No working init found.
		    Try passing init= option to kernel. See Linux Documentation/admin-guide/init.rst for guidance. ]---

+++ init/bootloader/u-boot/ubootdisk.sh
@@ -25,10 +25,24 @@ sudo mount ${dev_loop}p2 p2
 
 # Copy files
 #  p1: kernel + dtb
-sudo cp ${kernel_root_dir}/arch/arm64/boot/dts/arm/vexpress-v2f-1xv7-ca53x2.dtb p1/
-sudo cp ${kernel_root_dir}/arch/arm64/boot/Image{,.gz} p1/
+#   linux: aarch64 + vexpress_defconfig
+if [[ -e ${kernel_root_dir}/arch/arm64/boot/dts/arm/vexpress-v2f-1xv7-ca53x2.dtb ]]; then
+	sudo cp ${kernel_root_dir}/arch/arm64/boot/dts/arm/vexpress-v2f-1xv7-ca53x2.dtb p1/
+	sudo cp ${kernel_root_dir}/arch/arm64/boot/Image{,.gz} p1/
+#   linux: arm + sunxi_defconfig
+elif [[ -e ${kernel_root_dir}/arch/arm/boot/zImage ]]; then
+	sudo cp ${kernel_root_dir}/arch/arm/boot/zImage p1/
+	sudo cp ${kernel_root_dir}/arch/arm/boot/dts/allwinner/sun8i-h3-orangepi-pc.dtb p1/
+else
+	echo "WARNING: not found any kernel in ${kernel_root_dir}"
+	read -p "Press any key to continue."
+fi
 #  p2: rootfs
 
-sudo umount p1 p2
-rmdir p1 p2
-sudo losetup -d ${dev_loop}
+destroy() {
+	sudo umount p1 p2 || true
+	rmdir p1 p2 || true
+	sudo losetup -d ${dev_loop} || true
+}
+trap destroy EXIT
+
